requirements.txt
python-dotenv 
paramiko 
requests